Move noFilesFoundError to utils folder
