+++ lib/__tests__/ignore.test.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const NoFilesFoundError = require("../testUtils/noFilesFoundError");
+const NoFilesFoundError = require("../utils/noFilesFoundError");
 const path = require("path");
 const standalone = require("../standalone");
