+++ lib/__tests__/standalone.test.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const configBlockNoEmpty = require("./fixtures/config-block-no-empty");
-const NoFilesFoundError = require("../testUtils/noFilesFoundError");
+const NoFilesFoundError = require("../utils/noFilesFoundError");
 const path = require("path");
 const standalone = require("../standalone");
