+++ lib/standalone.js
@@ -14,7 +14,7 @@ const globby /*: Function*/ = require("globby");
 const hash = require("./utils/hash");
 const ignore = require("ignore");
 const needlessDisables /*: Function*/ = require("./needlessDisables");
-const NoFilesFoundError = require("./testUtils/noFilesFoundError");
+const NoFilesFoundError = require("./utils/noFilesFoundError");
 const path = require("path");
 const pify = require("pify");
 const pkg = require("../package.json");
